reading list updated

+++ src/app/modules/wishlist/wishlist.controller.ts
@@ -3,21 +3,23 @@ import * as WishlistService from './wishlist.service';
 import sendResponse from '../../utils/sendResponse';
 
 export const addToWishlistHandler = async (req: Request, res: Response): Promise<void> => {
-  const { book, user } = req.body;
-  const wishlist = await WishlistService.addToWishlist({ book, user });
+  const { book } = req.body;
+  const userId = req.user._id;
+  const wishlist = await WishlistService.addToWishlist({ book, user: userId });
 
   sendResponse(res, 201, 'Added to wishlist successfully', wishlist);
 };
 
 export const removeFromWishlistHandler = async (req: Request, res: Response): Promise<void> => {
-  const { book, user } = req.body;
-  const wishlist = await WishlistService.removeFromWishlist(book, user);
+  const { bookId } = req.params;
+  const userId = req.user._id;
+  const wishlist = await WishlistService.removeFromWishlist({ book: bookId, user: userId });
 
   sendResponse(res, 200, 'Successfully removed from wishlist', wishlist);
 };
 
 export const getWishlistHandler = async (req: Request, res: Response): Promise<void> => {
-  const { userId } = req.params;
+  const userId = req.user._id;
   const wishlist = await WishlistService.getWishlist(userId);
 
   sendResponse(res, 200, 'Wishlist retrieve successfully', wishlist);
